行业code由string转Integer报错，超出了Integer的范围

+++ src/main/java/com/pxene/pap/service/TencentAuditService.java
@@ -172,7 +172,7 @@ public class TencentAuditService extends AuditService
 			}
 			advertiserJson.addProperty("name", legalName);
 			advertiserJson.addProperty("homepage", siteUrl);
-			advertiserJson.addProperty("industry_id", Integer.parseInt(industryCode));
+			advertiserJson.addProperty("industry_id", Long.parseLong(industryCode));
 			JsonArray qualificationJsons = new JsonArray();
 			JsonObject qualificationJson = new JsonObject();
 			qualificationJson.addProperty("file_url", urlPrefix + qualificationPath);
